3~CDANetwork initialization and x86_64 initialization.

+++ kernel/kres/net/prot/core/ethernet.c
@@ -26,7 +26,7 @@ ethernet_send(
     return 0;
 
 fail:
-    return -1;
+    return (int) -1;
 }
